refact: move rbac handler to its own pkg
