+++ adapters/handlers/rest/rbac/handlers_authz_create_role_test.go
@@ -9,12 +9,13 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rest
+package rbacrest
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -27,11 +28,11 @@ import (
 	schemaMocks "github.com/weaviate/weaviate/usecases/schema/mocks"
 )
 
-func TestAddPermissionsSuccess(t *testing.T) {
+func TestCreateRoleSuccess(t *testing.T) {
 	type testCase struct {
 		name      string
 		principal *models.Principal
-		params    authz.AddPermissionsParams
+		params    authz.CreateRoleParams
 		// readCollection              bool
 		// readTenant                  bool
 		// readTenantWithoutCollection bool
@@ -41,9 +42,9 @@ func TestAddPermissionsSuccess(t *testing.T) {
 		{
 			name:      "all are *",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
-					Name: String("test"),
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
+					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
 							Action: String(authorization.CreateSchema),
@@ -55,8 +56,8 @@ func TestAddPermissionsSuccess(t *testing.T) {
 		{
 			name:      "collection checks",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
 					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
@@ -71,8 +72,8 @@ func TestAddPermissionsSuccess(t *testing.T) {
 		{
 			name:      "collection and tenant checks",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
 					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
@@ -89,8 +90,8 @@ func TestAddPermissionsSuccess(t *testing.T) {
 		{
 			name:      "* collections and tenant checks",
 			principal: &models.Principal{Username: "user1"},
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
 					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
@@ -111,14 +112,9 @@ func TestAddPermissionsSuccess(t *testing.T) {
 			schemaReader := schemaMocks.NewSchemaGetter(t)
 			logger, _ := test.NewNullLogger()
 
-			policies, err := conv.RolesToPolicies(&models.Role{
-				Name:        tt.params.Body.Name,
-				Permissions: tt.params.Body.Permissions,
-			})
-			require.Nil(t, err)
-
-			authorizer.On("Authorize", tt.principal, authorization.UPDATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(nil)
-			controller.On("UpsertRolesPermissions", policies).Return(nil)
+			authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+			controller.On("GetRoles", *tt.params.Body.Name).Return(map[string][]authorization.Policy{}, nil)
+			controller.On("UpsertRolesPermissions", mock.Anything).Return(nil)
 
 			// if tt.readCollection {
 			// 	schemaReader.On("ReadOnlyClass",
@@ -150,19 +146,49 @@ func TestAddPermissionsSuccess(t *testing.T) {
 				schemaReader: schemaReader,
 				logger:       logger,
 			}
-			res := h.addPermissions(tt.params, tt.principal)
-			parsed, ok := res.(*authz.AddPermissionsOK)
+			res := h.createRole(tt.params, tt.principal)
+			parsed, ok := res.(*authz.CreateRoleCreated)
 			assert.True(t, ok)
 			assert.NotNil(t, parsed)
 		})
 	}
 }
 
-func TestAddPermissionsBadRequest(t *testing.T) {
+func TestCreateRoleConflict(t *testing.T) {
+	authorizer := mocks.NewAuthorizer(t)
+	controller := mocks.NewController(t)
+	logger, _ := test.NewNullLogger()
+
+	principal := &models.Principal{Username: "user1"}
+	params := authz.CreateRoleParams{
+		Body: &models.Role{
+			Name: String("newRole"),
+			Permissions: []*models.Permission{
+				{
+					Action: String(authorization.CreateSchema),
+				},
+			},
+		},
+	}
+	authorizer.On("Authorize", principal, authorization.CREATE, authorization.Roles("newRole")[0]).Return(nil)
+	controller.On("GetRoles", *params.Body.Name).Return(map[string][]authorization.Policy{"newRole": {}}, nil)
+
+	h := &authZHandlers{
+		authorizer: authorizer,
+		controller: controller,
+		logger:     logger,
+	}
+	res := h.createRole(params, principal)
+	parsed, ok := res.(*authz.CreateRoleConflict)
+	assert.True(t, ok)
+	assert.Contains(t, parsed.Payload.Error[0].Message, fmt.Sprintf("role with name %s already exists", *params.Body.Name))
+}
+
+func TestCreateRoleBadRequest(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.AddPermissionsParams
-		principal     *models.Principal
+		params        authz.CreateRoleParams
+		upsertErr     error
 		expectedError string
 		// readCollection              bool
 		// readTenant                  bool
@@ -172,8 +198,8 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 	tests := []testCase{
 		{
 			name: "role name is required",
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
 					Name: String(""),
 					Permissions: []*models.Permission{
 						{
@@ -182,39 +208,36 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 					},
 				},
 			},
-			principal:     &models.Principal{Username: "user1"},
 			expectedError: "role name is required",
 		},
 		{
 			name: "role has to have at least 1 permission",
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
-					Name:        String("someName"),
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
+					Name:        String("newRole"),
 					Permissions: []*models.Permission{},
 				},
 			},
-			principal:     &models.Principal{Username: "user1"},
 			expectedError: "role has to have at least 1 permission",
 		},
 		{
-			name: "invalid action",
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
-					Name: String("someName"),
+			name: "invalid permission",
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
+					Name: String("someRole"),
 					Permissions: []*models.Permission{
 						{
-							Action: String("manage_somethingelse"),
+							Action: String("manage_something"),
 						},
 					},
 				},
 			},
-			principal:     &models.Principal{Username: "user1"},
 			expectedError: "invalid permission",
 		},
 		{
-			name: "update builtin role",
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
+			name: "cannot create role with the same name as builtin role",
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
 					Name: &authorization.BuiltInRoles[0],
 					Permissions: []*models.Permission{
 						{
@@ -223,17 +246,16 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 					},
 				},
 			},
-			principal:     &models.Principal{Username: "user1"},
-			expectedError: "you can not update builtin role",
+			expectedError: "you can not create role with the same name as builtin role",
 		},
 		// {
 		// 	name: "collection doesn't exist",
-		// 	params: authz.AddPermissionsParams{
-		// 		Body: authz.AddPermissionsBody{
+		// 	params: authz.CreateRoleParams{
+		// 		Body: &models.Role{
 		// 			Name: String("newRole"),
 		// 			Permissions: []*models.Permission{
 		// 				{
-		// 					Action:     String("manage_roles"),
+		// 					Action:     String(authorization.CreateSchema),
 		// 					Collection: String("ABC"),
 		// 				},
 		// 			},
@@ -244,12 +266,12 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 		// },
 		// {
 		// 	name: "tenant doesn't exist",
-		// 	params: authz.AddPermissionsParams{
-		// 		Body: authz.AddPermissionsBody{
+		// 	params: authz.CreateRoleParams{
+		// 		Body: &models.Role{
 		// 			Name: String("newRole"),
 		// 			Permissions: []*models.Permission{
 		// 				{
-		// 					Action:     String("manage_roles"),
+		// 					Action:     String(authorization.CreateSchema),
 		// 					Collection: String("ABC"),
 		// 					Tenant:     String("Tenant1"),
 		// 				},
@@ -261,8 +283,8 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 		// },
 		// {
 		// 	name: "tenant doesn't exist with * collection",
-		// 	params: authz.AddPermissionsParams{
-		// 		Body: authz.AddPermissionsBody{
+		// 	params: authz.CreateRoleParams{
+		// 		Body: &models.Role{
 		// 			Name: String("newRole"),
 		// 			Permissions: []*models.Permission{
 		// 				{
@@ -279,10 +301,16 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			controller := mocks.NewController(t)
 			authorizer := mocks.NewAuthorizer(t)
+			controller := mocks.NewController(t)
 			schemaReader := schemaMocks.NewSchemaGetter(t)
 			logger, _ := test.NewNullLogger()
+
+			if tt.expectedError == "" {
+				authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+				controller.On("GetRoles", *tt.params.Body.Name).Return(map[string][]authorization.Policy{}, nil)
+				controller.On("UpsertRolesPermissions", mock.Anything).Return(tt.upsertErr)
+			}
 
 			// if tt.readCollection {
 			// 	authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
@@ -312,14 +340,15 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 			// 		Physical: map[string]sharding.Physical{},
 			// 	})
 			// }
+
 			h := &authZHandlers{
-				controller:   controller,
 				authorizer:   authorizer,
+				controller:   controller,
 				schemaReader: schemaReader,
 				logger:       logger,
 			}
-			res := h.addPermissions(tt.params, tt.principal)
-			parsed, ok := res.(*authz.AddPermissionsBadRequest)
+			res := h.createRole(tt.params, nil)
+			parsed, ok := res.(*authz.CreateRoleBadRequest)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -329,10 +358,10 @@ func TestAddPermissionsBadRequest(t *testing.T) {
 	}
 }
 
-func TestAddPermissionsForbidden(t *testing.T) {
+func TestCreateRoleForbidden(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.AddPermissionsParams
+		params        authz.CreateRoleParams
 		principal     *models.Principal
 		authorizeErr  error
 		expectedError string
@@ -340,10 +369,10 @@ func TestAddPermissionsForbidden(t *testing.T) {
 
 	tests := []testCase{
 		{
-			name: "update some role",
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
-					Name: String("someRole"),
+			name: "authorization error",
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
+					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
 							Action: String(authorization.CreateSchema),
@@ -352,8 +381,8 @@ func TestAddPermissionsForbidden(t *testing.T) {
 				},
 			},
 			principal:     &models.Principal{Username: "user1"},
-			authorizeErr:  fmt.Errorf("some error from authZ"),
-			expectedError: "some error from authZ",
+			authorizeErr:  errors.New("authorization error"),
+			expectedError: "authorization error",
 		},
 	}
 
@@ -363,15 +392,15 @@ func TestAddPermissionsForbidden(t *testing.T) {
 			controller := mocks.NewController(t)
 			logger, _ := test.NewNullLogger()
 
-			authorizer.On("Authorize", tt.principal, authorization.UPDATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(tt.authorizeErr)
+			authorizer.On("Authorize", tt.principal, authorization.CREATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(tt.authorizeErr)
 
 			h := &authZHandlers{
 				authorizer: authorizer,
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.addPermissions(tt.params, tt.principal)
-			parsed, ok := res.(*authz.AddPermissionsForbidden)
+			res := h.createRole(tt.params, tt.principal)
+			parsed, ok := res.(*authz.CreateRoleForbidden)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -381,10 +410,10 @@ func TestAddPermissionsForbidden(t *testing.T) {
 	}
 }
 
-func TestAddPermissionsInternalServerError(t *testing.T) {
+func TestCreateRoleInternalServerError(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.AddPermissionsParams
+		params        authz.CreateRoleParams
 		principal     *models.Principal
 		upsertErr     error
 		expectedError string
@@ -392,10 +421,10 @@ func TestAddPermissionsInternalServerError(t *testing.T) {
 
 	tests := []testCase{
 		{
-			name: "update some role",
-			params: authz.AddPermissionsParams{
-				Body: authz.AddPermissionsBody{
-					Name: String("someRole"),
+			name: "upsert roles permissions error",
+			params: authz.CreateRoleParams{
+				Body: &models.Role{
+					Name: String("newRole"),
 					Permissions: []*models.Permission{
 						{
 							Action: String(authorization.CreateSchema),
@@ -404,8 +433,8 @@ func TestAddPermissionsInternalServerError(t *testing.T) {
 				},
 			},
 			principal:     &models.Principal{Username: "user1"},
-			upsertErr:     fmt.Errorf("some error from controller"),
-			expectedError: "some error from controller",
+			upsertErr:     errors.New("upsert error"),
+			expectedError: "upsert error",
 		},
 	}
 
@@ -415,16 +444,20 @@ func TestAddPermissionsInternalServerError(t *testing.T) {
 			controller := mocks.NewController(t)
 			logger, _ := test.NewNullLogger()
 
-			authorizer.On("Authorize", tt.principal, authorization.UPDATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(nil)
-			controller.On("UpsertRolesPermissions", mock.Anything).Return(tt.upsertErr)
+			policies, err := conv.RolesToPolicies(tt.params.Body)
+			require.Nil(t, err)
+
+			authorizer.On("Authorize", tt.principal, authorization.CREATE, authorization.Roles(*tt.params.Body.Name)[0]).Return(nil)
+			controller.On("GetRoles", *tt.params.Body.Name).Return(map[string][]authorization.Policy{}, nil)
+			controller.On("UpsertRolesPermissions", policies).Return(tt.upsertErr)
 
 			h := &authZHandlers{
 				authorizer: authorizer,
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.addPermissions(tt.params, tt.principal)
-			parsed, ok := res.(*authz.AddPermissionsInternalServerError)
+			res := h.createRole(tt.params, tt.principal)
+			parsed, ok := res.(*authz.CreateRoleInternalServerError)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -432,4 +465,8 @@ func TestAddPermissionsInternalServerError(t *testing.T) {
 			}
 		})
 	}
+}
+
+func String(s string) *string {
+	return &s
 }
