+++ adapters/handlers/rest/rbac/handlers_authz_get_role_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rest
+package rbacrest
 
 import (
 	"fmt"
@@ -17,6 +17,7 @@ import (
 
 	"github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 	"github.com/weaviate/weaviate/adapters/handlers/rest/operations/authz"
 	"github.com/weaviate/weaviate/entities/models"
 	"github.com/weaviate/weaviate/usecases/auth/authorization"
@@ -24,14 +25,14 @@ import (
 	"github.com/weaviate/weaviate/usecases/auth/authorization/mocks"
 )
 
-func TestGetRolesForUserSuccess(t *testing.T) {
+func TestGetRoleSuccess(t *testing.T) {
 	authorizer := mocks.NewAuthorizer(t)
 	controller := mocks.NewController(t)
 	logger, _ := test.NewNullLogger()
 
 	principal := &models.Principal{Username: "user1"}
-	params := authz.GetRolesForUserParams{
-		ID: "testUser",
+	params := authz.GetRoleParams{
+		ID: "testRole",
 	}
 
 	policies := []authorization.Policy{
@@ -42,52 +43,44 @@ func TestGetRolesForUserSuccess(t *testing.T) {
 		},
 	}
 
+	expectedPermissions, err := conv.PoliciesToPermission(policies...)
+	assert.Nil(t, err)
+
 	returnedPolices := map[string][]authorization.Policy{
 		"testRole": policies,
 	}
-
-	authorizer.On("Authorize", principal, authorization.READ, authorization.Roles("testRole")[0]).Return(nil)
-	controller.On("GetRolesForUser", params.ID).Return(returnedPolices, nil)
+	authorizer.On("Authorize", principal, authorization.READ, authorization.Roles(params.ID)[0]).Return(nil)
+	controller.On("GetRoles", params.ID).Return(returnedPolices, nil)
 
 	h := &authZHandlers{
 		authorizer: authorizer,
 		controller: controller,
 		logger:     logger,
 	}
-	res := h.getRolesForUser(params, principal)
-	parsed, ok := res.(*authz.GetRolesForUserOK)
+	res := h.getRole(params, principal)
+	parsed, ok := res.(*authz.GetRoleOK)
 	assert.True(t, ok)
 	assert.NotNil(t, parsed)
-
-	permissions, err := conv.PoliciesToPermission(policies...)
-	assert.Nil(t, err)
-
-	roles := []*models.Role{
-		{
-			Name:        String("testRole"),
-			Permissions: permissions,
-		},
-	}
-	expectedRoles := models.RolesListResponse(roles)
-	assert.Equal(t, expectedRoles, parsed.Payload)
+	assert.Equal(t, params.ID, *parsed.Payload.Name)
+	assert.Equal(t, expectedPermissions, parsed.Payload.Permissions)
 }
 
-func TestGetRolesForUserBadRequest(t *testing.T) {
+func TestGetRoleBadRequest(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.GetRolesForUserParams
+		params        authz.GetRoleParams
 		principal     *models.Principal
 		expectedError string
 	}
 
 	tests := []testCase{
 		{
-			name: "role name is required",
-			params: authz.GetRolesForUserParams{
+			name: "role id is required",
+			params: authz.GetRoleParams{
 				ID: "",
 			},
 			principal:     &models.Principal{Username: "user1"},
-			expectedError: "role name is required",
+			expectedError: "role id can not be empty",
 		},
 	}
 
@@ -100,8 +93,8 @@ func TestGetRolesForUserBadRequest(t *testing.T) {
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.getRolesForUser(tt.params, tt.principal)
-			parsed, ok := res.(*authz.GetRolesForUserBadRequest)
+			res := h.getRole(tt.params, tt.principal)
+			parsed, ok := res.(*authz.GetRoleBadRequest)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -111,10 +104,10 @@ func TestGetRolesForUserBadRequest(t *testing.T) {
 	}
 }
 
-func TestGetRolesForUserForbidden(t *testing.T) {
+func TestGetRoleForbidden(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.GetRolesForUserParams
+		params        authz.GetRoleParams
 		principal     *models.Principal
 		authorizeErr  error
 		expectedError string
@@ -122,9 +115,9 @@ func TestGetRolesForUserForbidden(t *testing.T) {
 
 	tests := []testCase{
 		{
-			name: "authorization error no access to role",
-			params: authz.GetRolesForUserParams{
-				ID: "testUser",
+			name: "authorization error",
+			params: authz.GetRoleParams{
+				ID: "testRole",
 			},
 			principal:     &models.Principal{Username: "user1"},
 			authorizeErr:  fmt.Errorf("authorization error"),
@@ -138,26 +131,15 @@ func TestGetRolesForUserForbidden(t *testing.T) {
 			controller := mocks.NewController(t)
 			logger, _ := test.NewNullLogger()
 
-			returnedPolices := map[string][]authorization.Policy{
-				"testRole": {
-					{
-						Resource: authorization.Collections("ABC")[0],
-						Verb:     authorization.READ,
-						Domain:   authorization.SchemaDomain,
-					},
-				},
-			}
-
-			authorizer.On("Authorize", tt.principal, authorization.READ, authorization.Roles("testRole")[0]).Return(tt.authorizeErr)
-			controller.On("GetRolesForUser", tt.params.ID).Return(returnedPolices, nil)
+			authorizer.On("Authorize", tt.principal, authorization.READ, authorization.Roles(tt.params.ID)[0]).Return(tt.authorizeErr)
 
 			h := &authZHandlers{
 				authorizer: authorizer,
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.getRolesForUser(tt.params, tt.principal)
-			parsed, ok := res.(*authz.GetRolesForUserForbidden)
+			res := h.getRole(tt.params, tt.principal)
+			parsed, ok := res.(*authz.GetRoleForbidden)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
@@ -167,10 +149,50 @@ func TestGetRolesForUserForbidden(t *testing.T) {
 	}
 }
 
-func TestGetRolesForUserInternalServerError(t *testing.T) {
+func TestGetRoleNotFound(t *testing.T) {
 	type testCase struct {
 		name          string
-		params        authz.GetRolesForUserParams
+		params        authz.GetRoleParams
+		principal     *models.Principal
+		expectedError string
+	}
+
+	tests := []testCase{
+		{
+			name: "role not found",
+			params: authz.GetRoleParams{
+				ID: "nonExistentRole",
+			},
+			principal:     &models.Principal{Username: "user1"},
+			expectedError: "role not found",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			authorizer := mocks.NewAuthorizer(t)
+			controller := mocks.NewController(t)
+			logger, _ := test.NewNullLogger()
+
+			authorizer.On("Authorize", tt.principal, authorization.READ, authorization.Roles(tt.params.ID)[0]).Return(nil)
+			controller.On("GetRoles", tt.params.ID).Return(map[string][]authorization.Policy{}, nil)
+
+			h := &authZHandlers{
+				authorizer: authorizer,
+				controller: controller,
+				logger:     logger,
+			}
+			res := h.getRole(tt.params, tt.principal)
+			_, ok := res.(*authz.GetRoleNotFound)
+			assert.True(t, ok)
+		})
+	}
+}
+
+func TestGetRoleInternalServerError(t *testing.T) {
+	type testCase struct {
+		name          string
+		params        authz.GetRoleParams
 		principal     *models.Principal
 		getRolesErr   error
 		expectedError string
@@ -178,9 +200,9 @@ func TestGetRolesForUserInternalServerError(t *testing.T) {
 
 	tests := []testCase{
 		{
-			name: "internal server error",
-			params: authz.GetRolesForUserParams{
-				ID: "testUser",
+			name: "internal server error from getting role",
+			params: authz.GetRoleParams{
+				ID: "testRole",
 			},
 			principal:     &models.Principal{Username: "user1"},
 			getRolesErr:   fmt.Errorf("internal server error"),
@@ -194,15 +216,28 @@ func TestGetRolesForUserInternalServerError(t *testing.T) {
 			controller := mocks.NewController(t)
 			logger, _ := test.NewNullLogger()
 
-			controller.On("GetRolesForUser", tt.params.ID).Return(nil, tt.getRolesErr)
+			policies := []authorization.Policy{
+				{
+					Resource: authorization.Collections("ABC")[0],
+					Verb:     authorization.READ,
+					Domain:   authorization.SchemaDomain,
+				},
+			}
+
+			returnedPolices := map[string][]authorization.Policy{
+				"testRole": policies,
+			}
+
+			authorizer.On("Authorize", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+			controller.On("GetRoles", tt.params.ID).Return(returnedPolices, tt.getRolesErr)
 
 			h := &authZHandlers{
 				authorizer: authorizer,
 				controller: controller,
 				logger:     logger,
 			}
-			res := h.getRolesForUser(tt.params, tt.principal)
-			parsed, ok := res.(*authz.GetRolesForUserInternalServerError)
+			res := h.getRole(tt.params, tt.principal)
+			parsed, ok := res.(*authz.GetRoleInternalServerError)
 			assert.True(t, ok)
 
 			if tt.expectedError != "" {
