+++ adapters/handlers/rest/rbac/handlers_authz_remove_permission_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package rest
+package rbacrest
 
 import (
 	"fmt"
